+++ mix/tasks/download_core_data.ex
@@ -1,32 +1,38 @@
 if File.exists?(Cldr.Config.download_data_dir()) do
   defmodule Mix.Tasks.Cldr.Download.CoreData do
     @moduledoc """
-    Downloads the pluralRanges.xml and timezones.xml files from CLDR
+    Downloads the measure.xml, pluralRanges.xml and timezones.xml files from CLDR
     """
 
     use Mix.Task
     require Logger
 
-    @shortdoc "Downloads the pluralRanges.xml and timezones.xml files from CLDR"
+    @shortdoc "Downloads measure.xml, pluralRanges.xml and timezones.xml files from CLDR"
 
     @url 'https://unicode.org/Public/cldr/'
+    @core 'core.zip'
+
     @plural_range 'common/supplemental/pluralRanges.xml'
     @timezones 'common/bcp47/timezone.xml'
-    @core 'core.zip'
+    @measurement_systems 'common/bcp47/measure.xml'
 
     @plural_ranges_output_file Path.join(Cldr.Config.download_data_dir(), "plural_ranges.xml")
     @timezones_output_file Path.join(Cldr.Config.download_data_dir(), "timezones.xml")
+    @measurement_systems_output_file Path.join(Cldr.Config.download_data_dir(), "measurement_systems.xml")
 
     def run(_) do
       with {:ok, content} <- download(url()),
            :ok <- File.write(to_string(@core), content),
-           {:ok, [{@timezones, timezones}, {@plural_range, plural_ranges}]} <-
-             :zip.unzip(@core, [:memory, {:file_list, [@plural_range, @timezones]}]) do
+           {:ok, [{@measurement_systems, measurement_systems}, {@timezones, timezones}, {@plural_range, plural_ranges}]} <-
+             :zip.unzip(@core, [:memory, {:file_list, [@measurement_systems, @plural_range, @timezones]}]) do
         File.write!(@plural_ranges_output_file, plural_ranges)
         Logger.info("Saved #{@plural_range} to #{@plural_ranges_output_file}")
 
         File.write!(@timezones_output_file, timezones)
         Logger.info("Saved #{@timezones} to #{@timezones_output_file}")
+
+        File.write!(@measurement_systems_output_file, measurement_systems)
+        Logger.info("Saved #{@measurement_systems} to #{@measurement_systems_output_file}")
 
         File.rm!(to_string(@core))
       else